# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API docs for eventhub",
        "title": "Eventhub Backend",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "paths": {
...
        "/communities": {
            "get": {
                "description": "Get all community with search and filter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "communities"
                ],
                "summary": "Get all community",
                "parameters": [
                    {
                        "type": "string",
                        "format": "category",
                        "description": "name category by category",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/communities/join": {
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Join to specified community",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "communities"
                ],
                "summary": "Join to community",
                "parameters": [
                    {
                        "description": "body to join community",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserCommunity"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/communities/leave": {
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Leave specified community",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "communities"
                ],
                "summary": "Leave community",
                "parameters": [
                    {
                        "description": "body to leave community",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserCommunity"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/communities/popular": {
            "get": {
                "description": "Get popular community",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "communities"
                ],
                "summary": "Get popular community",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/communities/{id}": {
            "get": {
                "description": "Get community detail",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "communities"
                ],
                "summary": "Get community detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Community ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/communities/{id}/events": {
            "get": {
                "description": "Get community event",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "communities"
                ],
                "summary": "Get community event",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Community ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/communities/{id}/members": {
            "get": {
                "description": "Get community member",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "communities"
                ],
                "summary": "Get community member",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Community ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "error"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "dto.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.UserCommunity": {
            "type": "object",
            "properties": {
                "community_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerToken": {
            "description": "Bearer Token used as identity for accessing backend",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}